Implement Parser, replace the old one
